Update BaseLayout.astro with Open Graph (OG) and Twitter meta tags for website type and URL
This commit adds the "og:type" and "og:url" meta tags to the BaseLayout.astro file, which are used by Open Graph protocol to identify the type of content as a website and provide its URL respectively. Additionally, it updates the Twitter meta tags with the new properties for title, description, image, and URL. This change improves the SEO and social sharing functionality of the application.